fixed subprocess
